/.well-known/core now returns the correct endpoint names
when using `char *url` the struct just have pointer to the
original string, which is removed from stack after the local
function ends.
By using char url[NANOCOAP_URL_MAX] we first reserve the memory
for the string statically, so that it will not be deleted after
the function ends.